Refactor more tests
