+++ tests/commands/create.ts
@@ -1,0 +1,46 @@
+import { expect } from 'chai';
+import { describe, it } from 'mocha';
+import * as fs from 'fs-extra';
+const { spawnSync } = require('child_process');
+
+import {
+  CLASP,
+} from '../constants';
+
+import {
+  cleanup,
+  setup,
+} from '../functions';
+
+import { LOG } from '../../src/utils';
+
+describe('Test clasp create function', () => {
+  before(setup);
+  it('should prompt for a project name correctly', () => {
+    spawnSync('rm', ['.clasp.json']);
+    const result = spawnSync(
+      CLASP, ['create'], { encoding: 'utf8' },
+    );
+    expect(result.stdout).to.contain(LOG.CLONE_SCRIPT_QUESTION);
+  });
+  it('should not prompt for project name', () => {
+    fs.writeFileSync('.clasp.json', '');
+    const result = spawnSync(
+      CLASP, ['create'], { encoding: 'utf8' },
+    );
+    expect(result.stderr).to.contain('Project file (.clasp.json) already exists.');
+  });
+  after(cleanup);
+});
+
+describe.skip('Test clasp create <title> function', () => {
+  before(setup);
+  it('should create a new project named <title> correctly', () => {
+    spawnSync('rm', ['.clasp.json']);
+    const result = spawnSync(
+      CLASP, ['create', 'myTitle'], { encoding: 'utf8' },
+    );
+    expect(result.stdout).to.contain('Created new script: https://script.google.com/d/');
+    expect(result.status).to.equal(0);
+  });
+});
