+++ tests/test.ts
@@ -13,10 +13,7 @@ import {
 } from './../src/utils';
 
 import {
-  backupSettings,
   cleanup,
-  restoreSettings,
-  rndStr,
   setup,
 } from './functions';
 
@@ -25,12 +22,8 @@ import {
   CLASP_SETTINGS,
   CLASP_USAGE,
   CLASP_PATHS,
-  CLIENT_CREDS,
-  FAKE_CLASPRC,
   IS_PR,
   SCRIPT_ID,
-  TEST_APPSSCRIPT_JSON,
-  TEST_CODE_JS,
 } from './constants';
 
 import {
@@ -65,99 +58,6 @@ describe('Test --help for each function', () => {
   it('should list --help', () => expectHelp('list', 'List App Scripts projects'));
   it('should apis --help', () => expectHelp('apis', 'List, enable, or disable APIs'));
   it('should help --help', () => expectHelp('help', 'Display help'));
-});
-
-describe('Test clasp list function', () => {
-  before(function () {
-    if (IS_PR) {
-      this.skip();
-    }
-  });
-  it('should list clasp projects correctly', () => {
-    const result = spawnSync(
-      CLASP, ['list'], { encoding: 'utf8' },
-    );
-    // Every project starts with this base URL, thus
-    // using clasp list should at least contain this
-    // in its output.
-    expect(result.stdout).to.contain('https://script.google.com/d/');
-    expect(result.status).to.equal(0);
-  });
-});
-
-describe('Test clasp create function', () => {
-  before(function () {
-    if (IS_PR) {
-      this.skip();
-    }
-  });
-  it('should prompt for a project name correctly', () => {
-    spawnSync('rm', ['.clasp.json']);
-    const result = spawnSync(
-      CLASP, ['create'], { encoding: 'utf8' },
-    );
-    expect(result.stdout).to.contain(LOG.CLONE_SCRIPT_QUESTION);
-  });
-  it('should not prompt for project name', () => {
-    fs.writeFileSync('.clasp.json', '');
-    const result = spawnSync(
-      CLASP, ['create'], { encoding: 'utf8' },
-    );
-    expect(result.stderr).to.contain('Project file (.clasp.json) already exists.');
-  });
-  after(cleanup);
-});
-
-describe.skip('Test clasp create <title> function', () => {
-  before(function () {
-    if (IS_PR) {
-      this.skip();
-    }
-  });
-  it('should create a new project named <title> correctly', () => {
-    spawnSync('rm', ['.clasp.json']);
-    const result = spawnSync(
-      CLASP, ['create', 'myTitle'], { encoding: 'utf8' },
-    );
-    expect(result.stdout).to.contain('Created new script: https://script.google.com/d/');
-    expect(result.status).to.equal(0);
-  });
-});
-
-describe('Test clasp clone <scriptId> function', () => {
-  before(function () {
-    if (IS_PR) {
-      this.skip();
-    }
-    setup();
-  });
-  it('should clone a project with scriptId correctly', () => {
-    cleanup();
-    const result = spawnSync(
-      CLASP, ['clone', SCRIPT_ID], { encoding: 'utf8' },
-    );
-    expect(result.stdout).to.contain('Cloned');
-    expect(result.stdout).to.contain('files.');
-    expect(result.status).to.equal(0);
-  });
-  it('should clone a project with scriptURL correctly', () => {
-    cleanup();
-    const result = spawnSync(
-      CLASP, ['clone', URL.SCRIPT(SCRIPT_ID)], { encoding: 'utf8' },
-    );
-    expect(result.stdout).to.contain('Cloned');
-    expect(result.stdout).to.contain('files.');
-    expect(result.status).to.equal(0);
-  });
-  it('should give an error on a non-existing project', () => {
-    fs.removeSync('./.clasp.json');
-    const result = spawnSync(
-      CLASP, ['clone', 'non-existing-project'], { encoding: 'utf8' },
-    );
-    expect(result.stderr).to.contain(ERROR.SCRIPT_ID);
-    expect(result.status).to.equal(1);
-  });
-  after(cleanup);
 });
 
 describe('Test extractScriptId function', () => {
@@ -212,50 +112,6 @@ describe('Test URL utils function', () => {
   });
 });
 
-describe('Test clasp deployments function', () => {
-  before(function () {
-    if (IS_PR) {
-      this.skip();
-    }
-    setup();
-  });
-  it('should list deployments correctly', () => {
-    const result = spawnSync(
-      CLASP, ['deployments'], { encoding: 'utf8' },
-    );
-    expect(result.stdout).to.contain('Deployment');
-    expect(result.status).to.equal(0);
-  });
-  after(cleanup);
-});
-
-describe('Test clasp deploy function', () => {
-  before(function () {
-    if (IS_PR) {
-      this.skip();
-    }
-    setup();
-  });
-  // Could fail to to maximum deployments (20)
-  // TODO: skip test if at maximum
-  it('should deploy correctly', () => {
-    const result = spawnSync(
-      CLASP, ['deploy'], { encoding: 'utf8' },
-    );
-    if (result.stderr) {
-      const err1 = 'Scripts may only have up to 20 versioned deployments at a time';
-      const err2 = 'Currently just one deployment can be created at a time';
-      const re = `(?:${err1}|${err2})`;
-      expect([result.stderr]).to.match(new RegExp(re));
-      expect(result.status).to.equal(1);
-    } else {
-      expect(result.stdout).to.contain('Created version ');
-      expect(result.status).to.equal(0);
-    }
-  });
-  after(cleanup);
-});
-
 describe('Test clasp version and versions function', () => {
   before(function () {
     if (IS_PR) {
@@ -286,41 +142,6 @@ describe('Test clasp version and versions function', () => {
     expect(result.stdout).to.contain('Versions');
     if (versionNumber) expect(result.stdout).to.contain(versionNumber + ' - ');
     expect(result.status).to.equal(0);
-  });
-  after(cleanup);
-});
-
-describe('Test clasp clone function', () => {
-  before(function () {
-    if (IS_PR) {
-      this.skip();
-    }
-    setup();
-  });
-  it('should prompt for which script to clone correctly', () => {
-    spawnSync('rm', ['.clasp.json']);
-    const result = spawnSync(
-      CLASP, ['clone'], { encoding: 'utf8' },
-    );
-    expect(result.stdout).to.contain('Clone which script?');
-  });
-  it('should prompt which project to clone and clone it', () => {
-    cleanup();
-    const result = spawnSync(
-      CLASP, ['clone'], { encoding: 'utf8', input: '\n'},
-    );
-    expect(result.stdout).to.contain('Clone which script?');
-    expect(result.stdout).to.contain('Cloned');
-    expect(result.stdout).to.contain('files.');
-    expect(result.status).to.equal(0);
-  });
-  it('should give an error if .clasp.json already exists', () => {
-    fs.writeFileSync('.clasp.json', '');
-    const result = spawnSync(
-      CLASP, ['clone'], { encoding: 'utf8' },
-    );
-    expect(result.stderr).to.contain('Project file (.clasp.json) already exists.');
-    expect(result.status).to.equal(1);
   });
   after(cleanup);
 });
@@ -450,153 +271,6 @@ describe('Test saveProject function from utils', () => {
     };
     expect(isSaved).to.not.equal(null);
   });
-});
-
-describe('Test clasp apis functions', () => {
-  before(function () {
-    if (IS_PR) {
-      this.skip();
-    }
-    setup();
-  });
-  it('should list apis correctly', () => {
-    const result = spawnSync(
-      CLASP, ['apis', 'list'], { encoding: 'utf8' },
-    );
-    expect(result.status).to.equal(0);
-    expect(result.stdout).to.contain('# Currently enabled APIs:');
-    expect(result.stdout).to.contain('# List of available APIs:');
-  });
-  it('should ask for an API when trying to enable', () => {
-    const result = spawnSync(
-      CLASP, ['apis', 'enable'], { encoding: 'utf8' },
-    );
-    expect(result.status).to.equal(1);
-    expect(result.stderr).to.contain('An API name is required.');
-  });
-  it('should enable sheets', () => {
-    const result = spawnSync(
-      CLASP, ['apis', 'enable', 'sheets'], { encoding: 'utf8' },
-    );
-    expect(result.status).to.equal(0);
-    expect(result.stdout).to.contain('Enabled sheets API.');
-  });
-  it('should give error message for non-existent API', () => {
-    const result = spawnSync(
-      CLASP, ['apis', 'enable', 'fakeApi'], { encoding: 'utf8' },
-    );
-    expect(result.status).to.equal(1);
-    expect(result.stderr).to.contain('API fakeApi doesn\'t exist. Try \'clasp apis enable sheets\'.');
-  });
-  it('should ask for an API when trying to disable', () => {
-    const result = spawnSync(
-      CLASP, ['apis', 'disable'], { encoding: 'utf8' },
-    );
-    expect(result.status).to.equal(1);
-    expect(result.stderr).to.contain('An API name is required.');
-  });
-  it('should disable apis correctly', () => {
-    const result = spawnSync(
-      CLASP, ['apis', 'disable', 'sheets'], { encoding: 'utf8' },
-    );
-    expect(result.status).to.equal(0);
-    expect(result.stdout).to.contain('Disabled sheets API.');
-  });
-  it('should show suggestions for using clasp apis', () => {
-    const result = spawnSync(
-      CLASP, ['apis'], { encoding: 'utf8' },
-    );
-    expect(result.status).to.equal(0);
-    expect(result.stdout).to.contain(`# Try these commands:
-- clasp apis list
-- clasp apis enable slides
-- clasp apis disable slides`);
-  });
-  it('should error with unknown subcommand', () => {
-    const result = spawnSync(
-      CLASP, ['apis', 'unknown'], { encoding: 'utf8' },
-    );
-    expect(result.status).to.equal(1);
-    expect(result.stderr).to.contain(`Unknown command`);
-  });
-  after(cleanup);
-});
-
-describe('Test clasp login function', () => {
-  before(function () {
-    if (IS_PR) {
-      this.skip();
-    }
-    setup();
-  });
-  beforeEach(backupSettings);
-  afterEach(restoreSettings);
-  it('should exit(0) with LOG.DEFAULT_CREDENTIALS for default login (no global or local rc)', () => {
-    if (fs.existsSync(CLASP_PATHS.rcGlobal)) fs.removeSync(CLASP_PATHS.rcGlobal);
-    if (fs.existsSync(CLASP_PATHS.rcLocal)) fs.removeSync(CLASP_PATHS.rcLocal);
-    const result = spawnSync(
-      CLASP, ['login', '--no-localhost'], { encoding: 'utf8' },
-    );
-    expect(result.stdout).to.contain(LOG.LOGIN(false));
-    expect(result.status).to.equal(0);
-  });
-  it('should ERROR.LOGGED_IN if global rc and no --creds option but continue to login', () => {
-    fs.writeFileSync(CLASP_PATHS.rcGlobal, FAKE_CLASPRC.token);
-    const result = spawnSync(
-      CLASP, ['login', '--no-localhost'], { encoding: 'utf8' },
-    );
-    fs.removeSync(CLASP_PATHS.rcGlobal);
-    expect(result.stderr).to.contain(ERROR.LOGGED_IN_GLOBAL);
-    expect(result.status).to.equal(0);
-  });
-  it('should exit(0) with ERROR.LOGGED_IN if local rc and --creds option', () => {
-    fs.writeFileSync(CLASP_PATHS.rcLocal, FAKE_CLASPRC.local);
-    const result = spawnSync(
-      CLASP, ['login', '--creds', `${CLASP_PATHS.clientCredsLocal}`, '--no-localhost'], { encoding: 'utf8' },
-    );
-    fs.removeSync(CLASP_PATHS.rcLocal);
-    expect(result.stderr).to.contain(ERROR.LOGGED_IN_LOCAL);
-    expect(result.status).to.equal(1);
-  });
-  it.skip('should exit(1) with ERROR.CREDENTIALS_DNE if --creds file does not exist', () => {
-    if (fs.existsSync(CLASP_PATHS.clientCredsLocal)) fs.removeSync(CLASP_PATHS.clientCredsLocal);
-    const result = spawnSync(
-      CLASP, ['login', '--creds', `${CLASP_PATHS.clientCredsLocal}`, '--no-localhost'], { encoding: 'utf8' },
-    );
-    expect(result.stderr).to.contain(ERROR.CREDENTIALS_DNE(CLASP_PATHS.clientCredsLocal));
-    expect(result.status).to.equal(1);
-  });
-  it.skip('should exit(1) with ERROR.BAD_CREDENTIALS_FILE if --creds file invalid', () => {
-    fs.writeFileSync(CLASP_PATHS.clientCredsLocal, CLIENT_CREDS.invalid);
-    const result = spawnSync(
-      CLASP, ['login', '--creds', `${CLASP_PATHS.clientCredsLocal}`, '--no-localhost'], { encoding: 'utf8' },
-    );
-    fs.removeSync(CLASP_PATHS.clientCredsLocal);
-    expect(result.stderr).to.contain(ERROR.BAD_CREDENTIALS_FILE);
-    expect(result.status).to.equal(1);
-  });
-  it.skip('should exit(0) with ERROR.BAD_CREDENTIALS_FILE if --creds file corrupt json', () => {
-    fs.writeFileSync(CLASP_PATHS.clientCredsLocal, rndStr());
-    const result = spawnSync(
-      CLASP, ['login', '--creds', `${CLASP_PATHS.clientCredsLocal}`, '--no-localhost'], { encoding: 'utf8' },
-    );
-    fs.removeSync(CLASP_PATHS.clientCredsLocal);
-    expect(result.stderr).to.contain(ERROR.BAD_CREDENTIALS_FILE);
-    expect(result.status).to.equal(1);
-  });
-  it('should exit(1) with LOG.CREDS_FROM_PROJECT if global rc and --creds file valid', () => {
-    if (fs.existsSync(CLASP_PATHS.rcLocal)) fs.removeSync(CLASP_PATHS.rcLocal);
-    fs.writeFileSync(CLASP_PATHS.rcGlobal, FAKE_CLASPRC.token);
-    fs.writeFileSync(CLASP_PATHS.clientCredsLocal, CLIENT_CREDS.fake);
-    const result = spawnSync(
-      CLASP, ['login', '--creds', `${CLASP_PATHS.clientCredsLocal}`, '--no-localhost'], { encoding: 'utf8' },
-    );
-    fs.removeSync(CLASP_PATHS.rcGlobal);
-    fs.removeSync(CLASP_PATHS.clientCredsLocal);
-    expect(result.stdout).to.contain(LOG.LOGIN(true));
-    expect(result.status).to.equal(1);
-  });
-  after(cleanup);
 });
 
 // Skipping for now because you still need to deploy function using GUI
